Revert "revert separate preprocess command"
This reverts commit 1660ef092399c20e188d6716285cba853d16eccf.
