+++ grants_tagger_light/preprocessing/__init__.py
@@ -1,8 +1,8 @@
 import typer
-
 from .preprocess_mesh import preprocess_mesh_cli
 
 preprocess_app = typer.Typer()
-preprocess_app.command("bioasq-mesh")(preprocess_mesh_cli)
-
-__all__ = ["preprocess_app"]
+preprocess_app.command(
+    "mesh",
+    context_settings={"allow_extra_args": True, "ignore_unknown_options": True},
+)(preprocess_mesh_cli)
