+++ grants_tagger_light/training/dataloaders/__init__.py
@@ -1,4 +1,3 @@
-from .mesh_json_loader import load_mesh_json
 from .multilabel_collator import MultilabelDataCollator
 
-__all__ = ["load_mesh_json", "MultilabelDataCollator"]
+__all__ = ["MultilabelDataCollator"]
